+++ frontend/src/metabase/query_builder/components/expressions/Editor/Tooltip/index.tsx
@@ -1,0 +1,1 @@
+export { Tooltip } from "./Tooltip";
